pyproject.toml
[project]
name = "prompt-injection-cv"
version = "0.1.0"
description = ""

authors = [
    { name = "Geevarghese George", email = "4496097+thatgeeman@users.noreply.github.com" },
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai (>=0.8.5,<0.9.0)",
    "notebook (>=7.4.4,<8.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "nbstripout (>=0.8.1,<0.9.0)",
    "google-genai (>=1.25.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = true
include = ["src/prompt_injection_cv"]
